+++ pkg/sdkv2/resources/fabric_connection_acceptance_test.go
@@ -6,8 +6,8 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/dbt-labs/terraform-provider-dbtcloud/pkg/dbt_cloud"
 	"github.com/dbt-labs/terraform-provider-dbtcloud/pkg/framework/acctest_helper"
+	"github.com/dbt-labs/terraform-provider-dbtcloud/pkg/helper"
 	"github.com/hashicorp/terraform-plugin-testing/helper/acctest"
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
 	"github.com/hashicorp/terraform-plugin-testing/terraform"
@@ -177,10 +177,15 @@ func testAccCheckDbtCloudFabricConnectionDestroy(s *terraform.State) error {
 		if rs.Type != "dbtcloud_fabric_connection" {
 			continue
 		}
-		projectId := strings.Split(rs.Primary.ID, dbt_cloud.ID_DELIMITER)[0]
-		connectionId := strings.Split(rs.Primary.ID, dbt_cloud.ID_DELIMITER)[1]
+		projectId, connectionId, err := helper.SplitIDToStrings(
+			rs.Primary.ID,
+			"dbtcloud_fabric_connection",
+		)
+		if err != nil {
+			return err
+		}
 
-		_, err := apiClient.GetConnection(connectionId, projectId)
+		_, err = apiClient.GetConnection(connectionId, projectId)
 		if err == nil {
 			return fmt.Errorf("Connection still exists")
 		}