+++ pkg/sdkv2/resources/project_artefacts_acceptance_test.go
@@ -6,8 +6,8 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/dbt-labs/terraform-provider-dbtcloud/pkg/dbt_cloud"
 	"github.com/dbt-labs/terraform-provider-dbtcloud/pkg/framework/acctest_helper"
+	"github.com/dbt-labs/terraform-provider-dbtcloud/pkg/helper"
 	"github.com/hashicorp/terraform-plugin-testing/helper/acctest"
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
 	"github.com/hashicorp/terraform-plugin-testing/terraform"
@@ -120,7 +120,13 @@ func testAccCheckDbtCloudProjectArtefactsExists(resource string) resource.TestCh
 		if err != nil {
 			return fmt.Errorf("Issue getting the client")
 		}
-		projectId := strings.Split(rs.Primary.ID, dbt_cloud.ID_DELIMITER)[0]
+		projectId, _, err := helper.SplitIDToStrings(
+			rs.Primary.ID,
+			"dbtcloud_project_artefacts",
+		)
+		if err != nil {
+			return err
+		}
 		project, err := apiClient.GetProject(projectId)
 		if err != nil {
 			return fmt.Errorf("Can't get project")
@@ -172,7 +178,13 @@ func testAccCheckDbtCloudProjectArtefactsDestroy(s *terraform.State) error {
 		if rs.Type != "dbtcloud_project_artefacts" {
 			continue
 		}
-		projectId := strings.Split(rs.Primary.ID, dbt_cloud.ID_DELIMITER)[0]
+		projectId, _, err := helper.SplitIDToStrings(
+			rs.Primary.ID,
+			"dbtcloud_project_artefacts",
+		)
+		if err != nil {
+			return err
+		}
 		project, err := apiClient.GetProject(projectId)
 		if project != nil {
 			return fmt.Errorf("Project still exists")