+++ pkg/sdkv2/resources/project_connection_acceptance_test.go
@@ -8,6 +8,7 @@ import (
 
 	"github.com/dbt-labs/terraform-provider-dbtcloud/pkg/dbt_cloud"
 	"github.com/dbt-labs/terraform-provider-dbtcloud/pkg/framework/acctest_helper"
+	"github.com/dbt-labs/terraform-provider-dbtcloud/pkg/helper"
 	"github.com/hashicorp/terraform-plugin-testing/helper/acctest"
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
 	"github.com/hashicorp/terraform-plugin-testing/terraform"
@@ -85,7 +86,13 @@ func testAccCheckDbtCloudProjectConnectionExists(resource string) resource.TestC
 		if err != nil {
 			return fmt.Errorf("Issue getting the client")
 		}
-		projectId := strings.Split(rs.Primary.ID, dbt_cloud.ID_DELIMITER)[0]
+		projectId, _, err := helper.SplitIDToStrings(
+			rs.Primary.ID,
+			"dbtcloud_project_connection",
+		)
+		if err != nil {
+			return err
+		}
 		project, err := apiClient.GetProject(projectId)
 		if err != nil {
 			return fmt.Errorf("Can't get project")
@@ -107,7 +114,12 @@ func testAccCheckDbtCloudProjectConnectionDestroy(s *terraform.State) error {
 		if rs.Type != "dbtcloud_project_connection" {
 			continue
 		}
-		projectId := strings.Split(rs.Primary.ID, dbt_cloud.ID_DELIMITER)[0]
+		projectId, _, err := helper.SplitIDToStrings(
+			rs.Primary.ID,
+			"dbtcloud_project_connection",
+		)
+		if err != nil {
+		
 		project, err := apiClient.GetProject(projectId)
 		if project != nil {
 			return fmt.Errorf("Project still exists")